requirements.txt
python-socketio==4.6.0
python-engineio==3.13.2
sanic==20.6.3
scikit-learn==0.23.2
scipy==1.5.2
numpy==1.23.1
pandas==1.0.5
requests==2.27.1
openpyxl==3.0.10
networkx==2.8.5